Allow variable field lengths size > 2 bytes long

+++ Dbarone.Net.Database/Core/Serialization/SerializationParams.cs
@@ -1,8 +1,19 @@
 public class ColumnSerializationInfo
 {
+    /// <summary>
+    /// The column name.
+    /// </summary>
     public string ColumnName { get; set; } = default!;
+    
+    /// <summary>
+    /// The column value.
+    /// </summary>
     public object? Value { get; set; }
-    public ushort Size { get; set; }
+    
+    /// <summary>
+    /// The size of the serialised value in bytes.
+    /// </summary>
+    public uint Size { get; set; }
 }
 
 /// <summary>
@@ -10,12 +21,43 @@ public class ColumnSerializationInfo
 /// </summary>
 public class SerializationParams
 {
-    public ushort TotalCount { get; set; }
-    public ushort TotalSize { get; set; }
-    public ushort FixedCount { get; set; }
+    /// <summary>
+    /// Total number of columns (max 255)
+    /// </summary>
+    public byte TotalCount { get; set; }
+    
+    /// <summary>
+    /// Total data size (uint).
+    /// </summary>
+    public uint TotalSize { get; set; }
+    
+    /// <summary>
+    /// Total number of fixed-length columns (max 255).
+    /// </summary>
+    public byte FixedCount { get; set; }
+    
+    /// <summary>
+    /// Total size of fixed-length columns (ushort).
+    /// </summary>
     public ushort FixedSize { get; set; }
+    
+    /// <summary>
+    /// The fixed length column information.
+    /// </summary>
     public List<ColumnSerializationInfo> FixedColumns { get; set; } = default!;
-    public ushort VariableCount { get; set; }
-    public ushort VariableSize { get; set; }
+    
+    /// <summary>
+    /// Total number of variable-length columns (max 255).
+    /// </summary>
+    public byte VariableCount { get; set; }
+    
+    /// <summary>
+    /// Size of variable-length data.
+    /// </summary>
+    public uint VariableSize { get; set; }
+    
+    /// <summary>
+    /// Variable length column information.
+    /// </summary>
     public List<ColumnSerializationInfo> VariableColumns { get; set; } = default!;
 }
